# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Go Copilot API
  description: |
    Authentication and Role-Based Access Control (RBAC) API.

    ## Authentication
    This API uses JWT Bearer tokens for authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Rate Limiting
    Sensitive endpoints are rate-limited. When limits are exceeded, a 429 response is returned with a Retry-After header.

    ## Permissions
    Access to resources is controlled via RBAC. Users are assigned roles, which contain permissions.
    Permission format: `resource:action` (e.g., `users:read`, `roles:create`)

    ## Common Flow Examples

    ### 1. User Registration and Login
    ```bash
    # Register a new user
    curl -X POST http://localhost:8080/api/v1/auth/register \
      -H "Content-Type: application/json" \
      -d '{"email": "user@example.com", "password": "SecurePass123!", "full_name": "John Doe"}'

    # Login to get tokens
    curl -X POST http://localhost:8080/api/v1/auth/login \
      -H "Content-Type: application/json" \
      -d '{"email": "user@example.com", "password": "SecurePass123!"}'
    ```

    ### 2. Making Authenticated Requests
    ```bash
    # Get current user profile
    curl -X GET http://localhost:8080/api/v1/auth/me \
      -H "Authorization: Bearer <access_token>"

    # List users (requires users:list permission)
    curl -X GET http://localhost:8080/api/v1/users \
...
    ```

    ### 3. Refreshing Tokens
...

...

...

...
servers:
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Exchange a valid refresh token for new access and refresh tokens
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Tokens refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
...
  /auth/logout-all:
    post:
      tags:
        - Authentication
      summary: Logout all sessions
      description: Revoke all refresh tokens for the current user
      operationId: logoutAll
      security:
        - bearerAuth: []
      responses:
        '204':
          description: All sessions logged out
        '401':
          description: Unauthorized
  /auth/forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: |
        Request a password reset email. Always returns success to prevent email enumeration.
      operationId: forgotPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset requested (always returns success)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
        '429':
          description: Rate limit exceeded
  /auth/reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset password using a valid reset token
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid or expired token, or weak password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user
      description: Get the authenticated user's profile with roles and permissions
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '401':
          description: Unauthorized
  /auth/sessions:
    get:
      tags:
        - Authentication
      summary: List active sessions
      description: Get all active sessions for the current user
      operationId: getSessions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of active sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionResponse'
        '401':
          description: Unauthorized
  /auth/sessions/{id}:
    delete:
      tags:
        - Authentication
      summary: Revoke session
      description: Revoke a specific session by ID
      operationId: revokeSession
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Session revoked
        '401':
          description: Unauthorized
        '404':
          description: Session not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
...
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: user@example.com
    ResetPasswordRequest:
      type: object
      required:
        - reset_token
        - new_password
      properties:
        reset_token:
          type: string
        new_password:
          type: string
          format: password
          minLength: 8
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserResponse'
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
    AuthUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        status:
          type: string
          enum: [active, inactive, banned]
        roles:
          type: array
          items:
            type: string
        permissions:
          type: array
          items:
            type: string
    SessionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        device_info:
          $ref: '#/components/schemas/DeviceResponse'
        ip_address:
          type: string
        created_at:
          type: string
          format: date-time
        last_used_at:
          type: string
          format: date-time
          nullable: true
        is_current:
          type: boolean
    DeviceResponse:
      type: object
      properties:
        user_agent:
          type: string
        platform:
          type: string
        browser:
          type: string
    ForgotPasswordResponse:
      type: object
      properties:
        message:
          type: string
          example: If an account exists with this email, a password reset link has been sent.
        expires_at:
          type: string
          format: date-time
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        avatar:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, banned]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string